array version. finished balanced node. ongoing: use array in processnode

+++ benchmarks/irregular_workload/lonestargpu/v5.5/apps/bfs/nodes.h
@@ -1,8 +1,8 @@
 struct Node
 {
   unsigned node;      // node to work on 
-  unsigned start;     // starting edge number
-  unsigned length;    // end edge number
+  unsigned short start;     // starting edge number
+  unsigned short length;    // end edge number
 //  Node *next;
 };
